Soln3FreeRTOS: Blink LED with delay time received over UART
This patch is inspired from challenge at  www.digiKey.in
(https://www.digikey.in/en/maker/projects/introduction-to
-rtos-solution-to-part-3-task-scheduling/8fbb9e0b0eed4279a2dd698f02ce125f)

Using FreeRTOS, create two separate tasks. One listens for
an integer over UART (from the Serial Monitor) and sets a
variable when it sees an integer. The other task blinks the
onboard LED (or other connected LED) at a rate specified by
that integer. In effect, you want to create a multi-threaded
system that allows for the user interface to run concurrently
with the control task (the blinking LED).