Soln4FreeRTOS: Memory Management Demo
This patch is inspired from challenge at  www.digiKey.in
(https://www.digikey.in/en/maker/projects/introduction-to-rtos-
solution-to-part-4-memory-management/6d4dfcaa1ff84f57a2098da8e6401d9c)

Using FreeRTOS, create two separate tasks. One listens for input over UART
(from the Serial Monitor). Upon receiving a newline character (‘\n’), the
task allocates a new section of heap memory (using pvPortMalloc()) and stores
the string up to the newline character in that section of heap. It then notifies
the second task that a message is ready.

The second task waits for notification from the first task. When it receives that
notification, it prints the message in heap memory to the Serial Monitor. Finally,
it deletes the allocated heap memory (using vPortFree()).