Soln7FreeRTOS: Semaphore challenge
This patch is inspired from challenge at  www.digiKey.in
(https://www.digikey.in/en/maker/projects/introduction-to-rtos
-solution-to-part-7-freertos-semaphore-example/51aa8660524c4daba38cba7c2f5baba7)

Starting with the code, add 2 semaphores and mutex to protect
and synchronize the circular buffer:

Challenge: use a mutex and counting semaphores to protect the shared buffer
so that each number (0 through 4) is printed exactly 3 times to the Serial
monitor (in any order). Do not use queues to do this!