Soln8FreeRTOS: Timer Challenge
This patch is inspired from challenge at  www.digiKey.in
(https://www.digikey.com/en/maker/projects/introduction-to
-rtos-solution-to-part-8-software-timers/0f64cf758da440a29476165a5b2e577e)

Your challenge is to create an auto-dim feature using the onboard LED.
We’ll pretend that the onboard LED (LED_BUILTIN) is the backlight to an LCD.
Create a task that echoes characters back to the serial terminal (as we’ve
done in previous challenges). When the first character is entered, the
onboard LED should turn on. It should stay on so long as characters are
being entered.

Extra: Enable gradual dimming and lightup.